ci(iast): move iast_patched_module to utils instead of conftest (#13009)
## Overview
This PR moves the `iast_patched_module` fixture from conftest.py to a
utils module to improve code organization and reusability in IAST tests.

## Motivation
- Reduces test configuration complexity by moving test utilities out of
conftest
- Makes the patched module functionality more accessible and reusable
across different test files
- Follows better practices by keeping fixtures focused and moving
utility code to dedicated modules

## Changes
- Moved `iast_patched_module` from conftest.py to a dedicated utils
module
- Updated related test files to use the new import location
- No functional changes, only code reorganization


## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
