+++ tests/appsec/iast/aspects/test_format_aspect_fixtures.py
@@ -15,7 +15,7 @@ from ddtrace.appsec._iast._taint_tracking._taint_objects import get_tainted_rang
 from ddtrace.appsec._iast._taint_tracking._taint_objects import taint_pyobject
 from tests.appsec.iast.aspects.aspect_utils import BaseReplacement
 from tests.appsec.iast.aspects.aspect_utils import create_taint_range_with_format
-from tests.appsec.iast.aspects.conftest import _iast_patched_module
+from tests.appsec.iast.iast_utils import _iast_patched_module
 from tests.utils import override_global_config
