+++ tests/appsec/iast/aspects/test_format_map_aspect_fixtures.py
@@ -8,7 +8,7 @@ import pytest
 from ddtrace.appsec._iast._taint_tracking import as_formatted_evidence
 from ddtrace.appsec._iast._taint_tracking import get_ranges
 from tests.appsec.iast.aspects.aspect_utils import BaseReplacement
-from tests.appsec.iast.aspects.conftest import _iast_patched_module
+from tests.appsec.iast.iast_utils import _iast_patched_module
 
 
 mod = _iast_patched_module("benchmarks.bm.iast_fixtures.str_methods")
