+++ tests/appsec/iast/iast_utils.py
@@ -1,13 +1,22 @@
+import importlib
 import re
+import types
 from typing import Optional
 from typing import Text
 import zlib
+
+from ddtrace.appsec._iast._ast.ast_patching import _should_iast_patch
+from ddtrace.appsec._iast._ast.ast_patching import astpatch_module
 
 
 # Check if the log contains "iast::" to raise an error if that’s the case BUT, if the logs contains
 # "iast::instrumentation::" or "iast::instrumentation::"
 # are valid logs
 IAST_VALID_LOG = re.compile(r"^iast::(?!instrumentation::|propagation::context::).*$")
+
+
+class IastTestException(Exception):
+    pass
 
 
 def get_line(label: Text, filename: Optional[Text] = None):
@@ -30,3 +39,20 @@ def get_line_and_hash(label: Text, vuln_type: Text, filename=None, fixed_line=No
     hash_value = zlib.crc32(rep.encode())
 
     return line, hash_value
+
+
+def _iast_patched_module_and_patched_source(module_name, new_module_object=False):
+    module = importlib.import_module(module_name)
+    module_path, patched_source = astpatch_module(module)
+    compiled_code = compile(patched_source, module_path, "exec")
+    module_changed = types.ModuleType(module_name) if new_module_object else module
+    exec(compiled_code, module_changed.__dict__)
+    return module_changed, patched_source
+
+
+def _iast_patched_module(module_name, new_module_object=False):
+    if _should_iast_patch(module_name):
+        module, _ = _iast_patched_module_and_patched_source(module_name, new_module_object)
+    else:
+        raise IastTestException(f"IAST Test Error: module {module_name} was excluded")
+    return module
