+++ tests/appsec/integrations/flask_tests/test_iast_langchain.py
@@ -2,8 +2,8 @@ import pytest
 
 from ddtrace.appsec._iast.constants import VULN_CMDI
 from ddtrace.internal.module import is_module_installed
-from tests.appsec.iast.aspects.conftest import _iast_patched_module
 from tests.appsec.iast.conftest import iast_context_defaults  # noqa: F401
+from tests.appsec.iast.iast_utils import _iast_patched_module
 from tests.appsec.iast.iast_utils import get_line_and_hash
 from tests.appsec.iast.taint_sinks.conftest import _get_span_report
 from tests.utils import override_env
